Move ZeroconfServiceInfo to service_info helpers
