+++ homeassistant/components/zeroconf/__init__.py
@@ -4,9 +4,8 @@ from __future__ import annotations
 
 import contextlib
 from contextlib import suppress
-from dataclasses import dataclass
 from fnmatch import translate
-from functools import lru_cache
+from functools import lru_cache, partial
 from ipaddress import IPv4Address, IPv6Address
 import logging
 import re
@@ -30,12 +29,19 @@ from homeassistant.const import (
     __version__,
 )
 from homeassistant.core import Event, HomeAssistant, callback
-from homeassistant.data_entry_flow import BaseServiceInfo
-from homeassistant.helpers import discovery_flow, instance_id
-import homeassistant.helpers.config_validation as cv
+from homeassistant.helpers import config_validation as cv, discovery_flow, instance_id
+from homeassistant.helpers.deprecation import (
+    DeprecatedConstant,
+    all_with_deprecated_constants,
+    check_if_deprecated_constant,
+    dir_with_deprecated_constants,
+)
 from homeassistant.helpers.discovery_flow import DiscoveryKey
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.network import NoURLAvailableError, get_url
+from homeassistant.helpers.service_info.zeroconf import (
+    ZeroconfServiceInfo as _ZeroconfServiceInfo,
+)
 from homeassistant.helpers.typing import ConfigType
 from homeassistant.loader import (
     HomeKitDiscoveredIntegration,
@@ -101,45 +107,11 @@ CONFIG_SCHEMA = vol.Schema(
     extra=vol.ALLOW_EXTRA,
 )
 
-
-@dataclass(slots=True)
-class ZeroconfServiceInfo(BaseServiceInfo):
-    """Prepared info from mDNS entries.
-
-    The ip_address is the most recently updated address
-    that is not a link local or unspecified address.
-
-    The ip_addresses are all addresses in order of most
-    recently updated to least recently updated.
-
-    The host is the string representation of the ip_address.
-
-    The addresses are the string representations of the
-    ip_addresses.
-
-    It is recommended to use the ip_address to determine
-    the address to connect to as it will be the most
-    recently updated address that is not a link local
-    or unspecified address.
-    """
-
-    ip_address: IPv4Address | IPv6Address
-    ip_addresses: list[IPv4Address | IPv6Address]
-    port: int | None
-    hostname: str
-    type: str
-    name: str
-    properties: dict[str, Any]
-
-    @property
-    def host(self) -> str:
-        """Return the host."""
-        return str(self.ip_address)
-
-    @property
-    def addresses(self) -> list[str]:
-        """Return the addresses."""
-        return [str(ip_address) for ip_address in self.ip_addresses]
+_DEPRECATED_HassioServiceInfo = DeprecatedConstant(
+    _ZeroconfServiceInfo,
+    "homeassistant.helpers.service_info.zeroconf.ZeroconfServiceInfo",
+    "2025.11",
+)
 
 
 @bind_hass
@@ -419,7 +391,7 @@ class ZeroconfDiscovery:
     def _async_dismiss_discoveries(self, name: str) -> None:
         """Dismiss all discoveries for the given name."""
         for flow in self.hass.config_entries.flow.async_progress_by_init_data_type(
-            ZeroconfServiceInfo,
+            _ZeroconfServiceInfo,
             lambda service_info: bool(service_info.name == name),
         ):
             self.hass.config_entries.flow.async_abort(flow["flow_id"])
@@ -595,7 +567,7 @@ def async_get_homekit_discovery(
     return None
 
 
-def info_from_service(service: AsyncServiceInfo) -> ZeroconfServiceInfo | None:
+def info_from_service(service: AsyncServiceInfo) -> _ZeroconfServiceInfo | None:
     """Return prepared info from mDNS entries."""
     # See https://ietf.org/rfc/rfc6763.html#section-6.4 and
     # https://ietf.org/rfc/rfc6763.html#section-6.5 for expected encodings
@@ -618,7 +590,7 @@ def info_from_service(service: AsyncServiceInfo) -> ZeroconfServiceInfo | None:
         assert service.server is not None, (
             "server cannot be none if there are addresses"
         )
-    return ZeroconfServiceInfo(
+    return _ZeroconfServiceInfo(
         ip_address=ip_address,
         ip_addresses=ip_addresses,
         port=service.port,
@@ -684,3 +656,11 @@ def _memorized_fnmatch(name: str, pattern: str) -> bool:
     since the devices will not change frequently
     """
     return bool(_compile_fnmatch(pattern).match(name))
+
+
+# These can be removed if no deprecated constant are in this module anymore
+__getattr__ = partial(check_if_deprecated_constant, module_globals=globals())
+__dir__ = partial(
+    dir_with_deprecated_constants, module_globals_keys=[*globals().keys()]
+)
+__all__ = all_with_deprecated_constants(globals())
