+++ homeassistant/config_entries.py
@@ -91,9 +91,9 @@ if TYPE_CHECKING:
     from .components.dhcp import DhcpServiceInfo
     from .components.ssdp import SsdpServiceInfo
     from .components.usb import UsbServiceInfo
-    from .components.zeroconf import ZeroconfServiceInfo
     from .helpers.service_info.hassio import HassioServiceInfo
     from .helpers.service_info.mqtt import MqttServiceInfo
+    from .helpers.service_info.zeroconf import ZeroconfServiceInfo
 
 
 _LOGGER = logging.getLogger(__name__)
